Identity service implemented for login and register API

+++ UserService/User.Domain/UserDto.cs
@@ -6,7 +6,7 @@ namespace User.Domain.Models
 
         public string Name { get; set; } = null!;
 
-        public int Mobile { get; set; }
+        public long Mobile { get; set; }
 
         public string Role { get; set; }
